part 4 ended

+++ controllers/blogs.js
@@ -1,7 +1,7 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
-
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async(request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -10,8 +10,7 @@ blogsRouter.get('/', async(request, response) => {
   
 blogsRouter.post('/', async(request, response, next) => {
     const body = request.body
-    
-    const user = await User.findById(body.user)
+    const user = request.user
     const blog = new Blog({
       title: body.title,
       author: body.author,
@@ -21,15 +20,25 @@ blogsRouter.post('/', async(request, response, next) => {
     })
 
     const savedBlog = await blog.save()
-    user.blogs = user.blogs.concat(savedBlog._id)
+    user.blogs = user.blogs.concat(savedBlog.id)
     await user.save()
     response.status(201).json(savedBlog)
 
   })
 
   blogsRouter.delete('/:id', async(request, response, next) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const blog = await Blog.findById(request.params.id)
+    const user = request.user
+    if ( blog.user.toString() === user._id.toString() ) {
+      await Blog.findByIdAndDelete(request.params.id)
+      const indexToDelete = await user.blogs.indexOf(request.params.id)
+      await user.blogs.splice(indexToDelete, 1)
+      await user.save()
+      response.status(204).end()
+    } else {
+      response.status(401).json({ error: 'Blog can only be deleted by its creator' })
+    }
+    
   })
 
   blogsRouter.get('/:id', async(request, response, next) => {
